requirements.txt
alembic==1.0.2
Babel==2.6.0
betterpath==0.2.2
blinker==1.4
Click==7.0
cymem==2.0.2
dominate==2.3.4
f==0.0.1
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
l==0.11.0
Mako==1.0.7
MarkupSafe==1.0
murmurhash==1.0.1
mysqlclient==1.3.14
numpy==1.15.3
pkg-resources==0.0.0
PyJWT==1.6.4
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.7
six==1.11.0
SQLAlchemy==1.2.13
vcversioner==2.16.0.0
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
wtforms-html5==0.2.2
zope.interface==4.6.0
